Dynamic product to invoice Added

+++ database/migrations/2019_08_06_045633_create_dealers_table.php
@@ -19,7 +19,7 @@ class CreateDealersTable extends Migration
             $table->string('contact')->nullable();
             $table->text('address')->nullable();
             $table->string('status')->nullable();
-            $table->integer('amount')->nullable();
+            $table->float('amount')->nullable();
             $table->timestamps();
         });
     }
